core: Remove DbFacade#isStoragePoolMasterUp(Guid)
The isStoragePoolMasterUp(Guid) method has no place in DbFacade, that's
supposed to be a general access point that provides the various DAOs.
Moreover, its implementation is awful, retrieving all of the pool's
domains from the database and iterating them one by one to find the
master domain in memory, instead of applying that filtering the query
itself.

This patch removes this useless method and its tests, and replaces it
with a method in IrsProxyData (the only place that used this call),
which is implemented properly, by retrieving only the correct domains
from the database, offering a slight performance benefit by the way.

Change-Id: I27b0ab4ae9b253ece29be97e9d9ae86ba9aefcb0
Signed-off-by: Allon Mureinik <amureini@redhat.com>