core: getEligibleVolumes performance
getEligibleVolumes traverses a list of GlusterVolumeEntitys and
filters it by a different predicate in each pass, resulting in a list
of GlusterVolumeEntitys that pass all the predicates.

This patch improves this method's performance by unifying all the
predicates to a single predicate using Predicate::and and then uses it
to filter the list. This change offers a slight performance
improvement as it saves the need to collect the results into a
different list each iteration.

Change-Id: I18b691366e4588263dcb99fe18d644f751458e69
Signed-off-by: Allon Mureinik <amureini@redhat.com>