core: StorageServerConnectionLunMapDaoImpl#getAll(String)
The getAll(String lunId) method was implemented naively be retrieving
all the rows from the table and then filtering them in memory.

This patch replaces this implementation by a proper stored procedure
implementation which is both easier to read, and potentially performs
better on large tables by leaving the filtering logic to the database
instead of transferring the entire table to the engine's memory and
having to filter there.

Change-Id: Ied9fbe3e427f37be604ae7d2fec58962f38cba22
Signed-off-by: Allon Mureinik <amureini@redhat.com>