restapi: BackendVmsResource template ID checking
Two separate places in this class compare IDs from business entities
(which are Guids) with IDs from API entities (which are Strings). Since
different types of objects cannot (or at least, should not) equal each
other in Java, these equality checks return false as long as at least
one of these values is not null, which is probably not what this code
intended.

This patch converts both values to Strings before comparing them. Using
ObjectUtils.equals hints to a concern that these values may be nulls, so
the null-safe Objects.toString is used to convert the Guids to Strings
instead of a straight-forward call to toString().

Change-Id: I94222e61ceebcd79168d62dd6449fb27730ea258
Signed-off-by: Allon Mureinik <amureini@redhat.com>