restapi: Set the reference to model in the anonymous module
Currently the anonymous module of the model doesn't link back to the
model, and the primitive types don't link back to the anonymous module.
The result of this is that calling the "getModel" method on primitive
types like "boolean" returns null, potentially causing null pointer
exceptions. This patch changes the "Model" class so that the anonymous
module links correctly to the model, and so that the primitive types
link correctly to the anonymous module.

Change-Id: I2475ec3ac9f84496bf389c49c2b21e1ea9b94b2a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>