restapi: Move schema type name generation to "SchemaNames"
Currently the method that generates a XML schema type name from an model
name is used only in order to generate XML schema, and it is located in
the "SchemaGenerator" class. In the future this method will also be
needed to generate the RSDL from the model. This patch moves it to the
"SchemaNames" class where it can be used from other places.

Change-Id: I01ceb285527f8a0eb00f70f09990fbcefe5b40b3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>