core: CommandBase - fix getParentParameters()
getParentParameters() is used to get the info for the command for which
the endAction() method should be executed when the created task ends.
When the parent command has a callback the parameters of the current command
should always be returned as the callback should be the one responsible
to call the endAction() when the child commands ends, otherwise the
AsyncTaskManager will execute it prematurely which is wrong.
This patch fixes this issue and adds some documentation to the method.

Change-Id: Ib49de599e4baf5f00b4e3117a4ef49cad6b91827
Signed-off-by: Liron Aravot <laravot@redhat.com>