core: HostUpgradeManager streaming API
Replace the usage of the outdated ListUtils in
HostUpgradeManager#getPackagesForCheckUpdate() with the modern streaming
API introduced in Java 8 as part of the effort to remove ListUtils from
the codebase completely.

In addition to cleaning up the code, this patch offers a slight
performance improvement by streaming both lists together and collecting
them into a single toSet() collector instead of adding all the packages
from the first list to a Set and then potentially rehahsing it when the
second (filtered) list is added.
A unit test was added to ensure that the proper functionallity is kept.

Change-Id: Ibb5b1d4197b1a353c7324fc5545e90580b5128b7
Signed-off-by: Allon Mureinik <amureini@redhat.com>