core: Remove ListUtils from SearchQuery
Removed the overly generic use of ListUtils in SearchQuery, in an
attempt to remove this class completely.

One instance (template searching) which actually needed a filter was
replaced with Java 8's new streaming API.
The second instance (VDS searching), which used the filter method but
had nothing to do with filtering was replaced by a good old for loop.

Change-Id: I7380a227d35d797387a17e850dadd4d2c842d465
Signed-off-by: Allon Mureinik <amureini@redhat.com>