engine: Storage validation in LSM flow
The patch I9f42f387781425d16f53a0e8a34d859365808ec0 introduced a change
that the disks validated in CreateAllSnapshotsFromVmCommand are all
the snappable disks.
This patch is reverting the mentioned above.

We need to check only the disks that the user selected.

LSM flow uses also CreateAllSnapshotsFromVmCommand.
In this case, the disks are already validated in the LSM command,
no need to check them again.

Change-Id: I0bb0aa45f54824ed93e299827c3f38c0905bb3cd
Bug-Url: https://bugzilla.redhat.com/1282957
Bug-Url: https://bugzilla.redhat.com/1287025
Signed-off-by: Fred Rolland <frolland@redhat.com>