webadmin: isOptionEnabledOnVolume filters
isOptionEnabledOnVolume receives a list of filters, but is always
called with a singlton list containing a single element. Taking into
consideration that this code was not touched for over a year, the method
can be simplified to receive a single filter, in order to make it easier
to read.

In the future, if additional filters are required, this patch could
always be reverted.

Change-Id: Ide95e55e72f082e3969857b8e3ba13d668316130
Signed-off-by: Allon Mureinik <amureini@redhat.com>