core: Video RAM size settings reworked
Video RAM sizes are currently set in a wrong way.  Especially one of the
most important parameters, `vgamem', is not set at all.  That may
prevent setups with high resolution monitors from working.

This patch tries to set more reasonable video RAM sizes.  It replaces
the old video RAM size computation (including removal of some unused
related code) with a new one, based on guest OS.  The guest OS specifics
are set in osinfo.

We try to keep the computation as simple as possible, for the following
reasons:
1. Nobody can say what the perfect values are.
2. Things are likely to change in QEMU in future.
3. We try to be reasonably conservative not to break current setups.

Note that recent VDSM version is needed to set `vgamem' value (it's
ignored in older versions).

As we can't be sure about all older versions of VDSM, libvirt and QEMU,
we send the same values as before to clusters of older versions.  These
may be not the best values but it's the safest approach.  We may provide
means to override the memory values via osinfo in a followup patch.

Change-Id: I96ab9d40621ad12ec2505a2de75e3ff5af2f59d0
Bug-Url: https://bugzilla.redhat.com/1275539
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>