core: Adding CommandStatus indicating the endAction() was executed
Currently when command async operations execution ends its status is set to
SUCCEEDED/FAILED (either by the callback or in CommandBase if it has no
callback.) - The problem is that this status doesn't indicate that the
endAction() execution was ended - this may cause the parent to determine
that a child command was ended prematurely.

For commands that has callbacks the CommandEntity.isCallbackNotified()
flag is set after the onSucceeded()/onFailed() method are called and may
be used as an indication that the endAction() method was executed - this
may be confusing (as one need to remember to check the command status
and this flag as well). Furthremore, this flag is currently not uses for
AsyncTaskManager managed commands.

This patch adds new command statuses, ENDED_SUCCESSFULLY and
ENDED_WITH_FAILURE that indicates that the endAction() method logic was
executed, in a further patch we can look into remove the isCallbackNotified()
flag.

Change-Id: Ib1ae7b2c247ce5343352cbdd70e8fa060d25f7dd
Signed-off-by: Liron Aravot <laravot@redhat.com>