core: CommandBase - Parent may run the child commands endAction()
Command Coordinator commands may track commands with tasks and want to
execute their endSuccesfully()/endWithFailure() methods independently
and not when the task execution ends.

If the parent command has a callback (as this change is introduced to be
used with the CoCo framework only), the set value in VdcActionParametersBase
will be checked to determine how to proceed, if needed the command endAction()
method will be executed otherwise it'll be logged that the parent
command should end the child command.

Change-Id: I84820416344dd87626eb0805c3c273ee47b95cb7
Signed-off-by: Liron Aravot <laravot@redhat.com>