core: set the shouldBeEndedByParent flag when needed
Currently the shouldBeEndedByParent flag isn't used by any command, the
problem is that the commands updated in this patch currently rely on the
fact that the child command endAction() method will be executed on its
completion by its callback and not by the parent, which collides with
the shouldBeEndedByParent default value (true).
This patch makes it clear that the child command endAction() should run
when its execution is over by setting this flag to false on those
scenarios.

Change-Id: I5b03c71a1808a55b658889d5b6c79a257d0a4179
Signed-off-by: Liron Aravot <laravot@redhat.com>