restapi: Make "JavaClassBuffer" a member instead of component
The class "JavaClassBuffer" used to generate Java code used to be a CDI
component, because it was created in one place and them used in other
places (to add "import" statements). That isn't required now because the
"JavaClassName" class contains all the required information, and can be
passed between components. This patch changes the "JavaClassBuffer" so
that it is created where needed and no longer an application scoped CDI
component.

Change-Id: Ic01813f186e670d5bc66035e1358e9614a85d618
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>