restapi: Add "--version-prefix" option to metamodel tools
This patch adds support for the "--version-prefix" option to the
metamodel tool. This option will be used in later patches to geneate
Java classes from the model with a prefix in the name. For example the
name of the class generated for the "Vm" type will be "V4Vm" in version
4 of the API and "V5Vm" in version 3 of the API. This way both classes
can be used in the server, without having to use long fully qualified
names.

Change-Id: Id1fd275cdac58c60652e0f62115a320e8cc1f2df
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>