core: Remove unused Tag methods and classes
Removed the unused methods getAllUserGroupTagsWithIds(String),
getAllForVdsWithIds(String) and
getVmPoolTagsByVmPoolIdAndAdElementId(Guid, Guid) from TagDao and its
implementation, as well as the backing stored procedures that serve
these methods.

getVmPoolTagsByVmPoolIdAndAdElementId was the only place in the codebase
to use TagsVmPoolMap, so it was also removed. Once that was removed,
TagsVmPoolMapId also became redundant, and was removed as well.

Change-Id: I120f3c6f4f89e85887593d5ff87c7f529551e257
Signed-off-by: Allon Mureinik <amureini@redhat.com>