core: rename canDoAction to validate
canDoAction/CanDoAction appears all over the code.
This is misleading since the more proper term is validate/Validate.

This patch replaces all *canDoAction*/*CanDoAction* with
*validate*/*Validate* all over the code :

In class interface names
In variables
In methods
In log messages

validate was chosen since in phases names that the Task Manager displays
for a running command the canDoAction step is called VALIDATING :

public enum StepEnum {
    VALIDATING,
    EXECUTING,
    FINALIZING,
    REBALANCING_VOLUME,
    REMOVING_BRICKS,
    UNKNOWN;
}

This patch makes the code more clear and readable and improve the
clearness of our log messages.

Work was mainly done using sed

Change-Id: I0074620ba72a8f4768ffdb06122b6345f39f780d
Signed-off-by: emesika <emesika@redhat.com>