core: CoCoAsyncTaskHelper - get parent command id from the parent parameters
When choosing parent parameters for the tasks the parent command id should be
derived from those parameters, otherwise we can get into situation where
the parentCommandId will be Guid.Empty not matching the selected parent
parameters.

This situation leads to the tasks to be considered as of different commands which
leads to the endAction() method to be called multiple times when the parent command
creates tasks. An example would be the usage of the memory builders, when creating
a vm snapshot with memory the endAction() will be executed multiple times, one time
for the disks tasks and one time for the memory tasks.

Change-Id: I0cacc023e855c8e9651e7134234759288b93e897
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1293663
Signed-off-by: Liron Aravot <laravot@redhat.com>