engine: add 'host_device' column to 'vm_device' table
Since the 'hostdev' property was already reported by the vdsm before this patch,
after adding the 'host_device' column to 'vm_device' table, 
already existing vms will continue report the same hash (since in the vdsm side,
no change was done).
It will cause the 'vm-device' table not to be updated with the 'hostdev' value
(until an actual change will be done to the vm).
In order to solve this, the patch nullifies all the hashes of vms with plugged,
passthrough vnic/s.

Change-Id: I0c5c36365f98f3ed4243562b474b846866d8ba62
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>