core: hosted-engine: Lock the sd import exclusively
Add an exclusive lock on the hosted engine domain import to prevent it
from interleaving. There is a race where the domain might be added but
the TX isn't closed yet and another import can enter and pass the
command validation.

Small clean-up made to remove the redundant getBackend() and
 runVdsCommand(..)  methods.

Change-Id: I2b67db78433f3c0847e118ccb279d63a65986b34
Bug-Url: https://bugzilla.redhat.com/1293928
Bug-Url: https://bugzilla.redhat.com/1290518
Signed-off-by: Roy Golan <rgolan@redhat.com>