engine: Add default graphics devices if absent in OVF
If OVF was created on pre-3.6 version of the Engine, graphics devices
will be absent in it. After restoration under Engine 3.6, the VM without
graphics devices will fail to open the console or even fail to run (on
recent versions of VDSM).

This patch creates graphics devices, if absent after reading an OVF.
The device type (protocol) is guessed from default display type of the
VM.

Change-Id: I596cbb9af4573fd25d1072c911227fb3bd759e0b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1293299
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>