core: compare custom properties against template on add vm
We should require permission for action-group of type
ActionGroup.CHANGE_VM_CUSTOM_PROPERTIES on add-VM only if the
custom properties of the VM to add are different than the ones
defined in the template it is based on.

Currently, since we only check whether the VM is defined with
custom properties or not and by default the VM gets the custom
properties of its template, users cannot create a VM that is
based on template which is defined with custom properties
without having these extra permissions.

Change-Id: I1c3bb293085cc26ad93f470c4d7e151cc2af08f0
Bug-Url: https://bugzilla.redhat.com/1284472
Signed-off-by: Arik Hadas <ahadas@redhat.com>