core: fix import vm from vmware as clone
This patch fixes the functionality of import-as-clone from VMware.
The problem was that on import-as-clone we generate new ID, but
in the ImportFromExternalProviderCommand we never used the generated
ID.

The solution is composed of 2 parts:

1. To move the getters for VM-ID that reside in ImportVmCommand and
knows to return the right ID, also for the import-as-clone case, to
ImportVmCommandBase so they will be shared with the command
ImportFromExternalProviderCommand as well.

2. We need the parameters to store the ID of the VM in two fields:
getVm().getId() and getVmId(). Instead of setting it manually, we
now call the constructor of ImportVmParameters that does the trick.

Change-Id: I32dfe891c6168d3c1d522e2588ae9e02d82b9de3
https://bugzilla.redhat.com/1294501
Signed-off-by: Arik Hadas <ahadas@redhat.com>